tokenize attributes and identify closing tags
Tokenize now keeps track of which tags can be closed and ignores closing
tag markers that are impossible (they just get treated as text).

This was partly so that we could have pretty attribute markup, with just
a ] as a closing tag, and partly because it would eventually become a
problem anyway.